requirements.txt
pyrogram
tgcrypto
numpy
python_dotenv
firebase_admin
